+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -202,4 +202,25 @@ public class FileUtilTest {
 			Console.log(file.getPath());
 		}
 	}
+	
+	@Test
+	public void getParentTest() {
+		File parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 0);
+		Assert.assertEquals(FileUtil.file("d:\\aaa\\bbb\\cc\\ddd"), parent);
+		
+		parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 1);
+		Assert.assertEquals(FileUtil.file("d:\\aaa\\bbb\\cc"), parent);
+		
+		parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 2);
+		Assert.assertEquals(FileUtil.file("d:\\aaa\\bbb"), parent);
+		
+		parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 4);
+		Assert.assertEquals(FileUtil.file("d:\\"), parent);
+		
+		parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 5);
+		Assert.assertNull(parent);
+		
+		parent = FileUtil.getParent(FileUtil.file("d:/aaa/bbb/cc/ddd"), 10);
+		Assert.assertNull(parent);
+	}
 }